LPS-112693 Add test
This test sets up the following scenario:
1. A calendar event is created on 2016 Jan 01 that repeats until X date
2. The event on 2016 Jan 01 is deleted
3. The event on 2016 Jan 02 is modified so that it repeats until 2016
Jan 03

The entire series now starts from 2016 Jan 02 and ends on 2016 Jan 03.
Because of the modification in Step 3, this scenario now consists of two
CalendarBookings which share the same recurringCalendarBookingId:
1. A CalendarBooking on 2016 Jan 01 that repeats until 2016 Jan 01
2. A CalendarBooking on 2016 Jan 02 that repeats until 2016 Jan 03

Retrieving the last CalendarBooking instance for the 2016 Jan 01
CalendarBooking expands into an empty list. An
ArrayIndexOutOfBoundsException occurs as a result due to a .size() - 1
operation being performed on the empty list.